+++ app/sdks/server-nodejs/lib/client.js
@@ -55,19 +55,6 @@ class Client {
         return this;
     }
 
-    /**
-     * Set Mode
-     *
-     * @param string value
-     *
-     * @return self
-     */
-    setMode(value) {
-        this.addHeader('X-Appwrite-Mode', value);
-
-        return this;
-    }
-
     /***
      * @param bool status
      * @return this